Initial failed attempt with faststream
Main reason that it's not possible to bootstrap the project with
FastStream is because of the double broker.subscriber and publish
decorator. Why does it need both the subscriber and publisher to produce
streams? In our case, the application starts with KlinesConnect, so that
means we'd need to decorate a class and the producer/publisher is not
triggered until there is incoming websocket data