Remove partition separation
Kafka partitions need to be ordered, otherwise it will fail to send.
Having the partitions in the DB is not useful, because when initializing the app, you can't stream data using a partition e.g. 365, it must start from 0.
However, we do need the to produce because kakfa streams are faster than MongoDB on streams.