Provide update_required with Kafka streams (#14)
* Use new simplied balance estimation endpoints to fix timeout

* Autotrade signals reception fixes

* Optimize docker image

* Temporarily remove pyspark due to server space limits in prod

* Replace all instances of update_required with new kafka streams systems

This new system adds the usually update_required signal, as a Kafka message. In addition it appends an event name of the format "EVENT_TYPE_ACTION" all in uppercase that describes where this is produced. To be refactored in an Enum in the future. This should fix #14, reduce duplication of autotrade signals.