Store kline timestamps as integer
Datetime objects will not work, because precision of such objects are not enough, timestamps provided are much more precise.
Usually they don't come with decimals, so integers will work fine.
The reason why we need to store it as integers is so that MongoDB can order correctly with this number.