Fix missing module and message validation and serialization (#17)
* Fix missing module

* Corrections to bb spreads

* Replace volatility with top bollinguer band

After analyzing telegram signals, it seems votality in terms of logarithmic standard deviation doesn't correlate in any way to the spread of Bollinguer bands. Therefore, directly use bb bands for filtering signals

* Fix prod not emitting signals

For some reason, locally we are getting signals but not in prod. This is an attempt to get signals in prod.