Test remove reliance of own klines by sequentially queuing kline data to avoid weight limits
Stored MongoDB kline data is too inconsistent and unreliable. Candles are often lagging behind and do not trigger the signals when they should. This is caused mainly by the lack of capacity in the server to store humongous amount of data (1m interval candlesticks for over 277 assets with over 500 limit datapoints at least, this is usually GBs of data), which will drive up costs to $100 a month, that would be sort of level needed to store minute candlesticks, to consistenly provide accurate signals. But we can't afford it at the moment.

This approach will remove reliance on stored kline data, by increasing the buffer of the signals (kafka simply would be emitting 5m klines sequentially, so it does not hit rate limits). Of course, this is assuming 5m kline, which is a good middle point:
- Too low and we don't get enough websocket updates
- Too high and we'll miss too many spike signals

Also we assume we are only listening to a small amount of assets (at the time of writing ~277), so that means there's not a huge amount of requests made to the Binance API: 277/5 = 55 requests per minute, to break-even, we could drag it a little bit into maybe 10min (there are no kline intervals between 15m and 5m), so that means 27-28 requests per munite.